Fix: Price is f64

+++ ejercicios/rest-workshop/ej2/src/model/model.go
@@ -1,9 +1,9 @@
 package model
 
 type Item struct {
-	Name  string `json:"name" binding:"required"`
-	ID    string `json:"id" binding:"required"`
-	Price int    `json:"price" binding:"required"`
+	Name  string  `json:"name" binding:"required"`
+	ID    string  `json:"id" binding:"required"`
+	Price float64 `json:"price" binding:"required"`
 }
 
 type Error struct {
